undo the change of moving scala api
