+++ akka-actor-tests/src/test/scala/akka/actor/FSMActorSpec.scala
@@ -6,7 +6,6 @@ package akka.actor
 
 import language.postfixOps
 import akka.testkit._
-import akka.testkit.scaladsl.TestKit
 import scala.concurrent.duration._
 import akka.event._
 import com.typesafe.config.ConfigFactory
