+++ akka-actor-tests/src/test/scala/akka/actor/FSMTimingSpec.scala
@@ -4,8 +4,6 @@
 
 package akka.actor
 
-import akka.testkit.scaladsl.TestKit
-
 import language.postfixOps
 
 import akka.testkit._
