+++ akka-actor-tests/src/test/scala/akka/dispatch/DispatcherShutdownSpec.scala
@@ -5,7 +5,7 @@
 package akka.dispatch
 
 import akka.actor.ActorSystem
-import akka.testkit.scaladsl.TestKit
+import akka.testkit.TestKit
 import java.lang.management.ManagementFactory
 import org.scalatest.{ Matchers, WordSpec }
 import scala.concurrent.{ Await, Future }
