+++ akka-camel/src/test/scala/akka/camel/ProducerFeatureTest.scala
@@ -17,7 +17,6 @@ import scala.concurrent.duration._
 import akka.util.Timeout
 import org.scalatest.Matchers
 import akka.testkit._
-import akka.testkit.scaladsl.TestKit
 import akka.actor.Status.Failure
 
 /**
