+++ akka-camel/src/test/scala/akka/camel/TestSupport.scala
@@ -15,8 +15,7 @@ import scala.reflect.ClassTag
 import akka.actor.{ ActorRef, Props, ActorSystem, Actor }
 import scala.concurrent.Await
 import akka.util.Timeout
-import akka.testkit.AkkaSpec
-import akka.testkit.scaladsl.TestKit
+import akka.testkit.{ TestKit, AkkaSpec }
 
 private[camel] object TestSupport {
   def start(actor: ⇒ Actor, name: String)(implicit system: ActorSystem, timeout: Timeout): ActorRef =
