+++ akka-cluster-tools/src/test/scala/akka/cluster/pubsub/DistributedPubSubMediatorRouterSpec.scala
@@ -1,7 +1,6 @@
 package akka.cluster.pubsub
 
 import akka.testkit._
-import akka.testkit.scaladsl.TestKit
 import akka.routing.{ ConsistentHashingRoutingLogic, RouterEnvelope }
 import org.scalatest.WordSpecLike
 import akka.actor.{ DeadLetter, ActorRef }
