+++ akka-cluster-tools/src/test/scala/akka/cluster/singleton/ClusterSingletonProxySpec.scala
@@ -4,8 +4,7 @@
 package akka.cluster.singleton
 
 import org.scalatest.{ BeforeAndAfterAll, Matchers, WordSpecLike }
-import akka.testkit.scaladsl.TestKit
-import akka.testkit.TestProbe
+import akka.testkit.{ TestProbe, TestKit }
 import akka.actor._
 import com.typesafe.config.ConfigFactory
 import akka.cluster.Cluster
