+++ akka-cluster/src/test/scala/akka/cluster/DowningProviderSpec.scala
@@ -7,9 +7,8 @@ import java.util.concurrent.atomic.AtomicBoolean
 
 import akka.ConfigurationException
 import akka.actor.{ ActorSystem, Props }
-import akka.testkit.scaladsl.TestKit._
+import akka.testkit.TestKit.{ awaitCond, shutdownActorSystem }
 import com.typesafe.config.ConfigFactory
-import org.scalatest.concurrent.Futures
 import org.scalatest.{ Matchers, WordSpec }
 
 import scala.concurrent.duration._
