+++ akka-contrib/src/test/scala/akka/contrib/pattern/AggregatorSpec.scala
@@ -3,8 +3,7 @@
  */
 package akka.contrib.pattern
 
-import akka.testkit.ImplicitSender
-import akka.testkit.scaladsl.TestKit
+import akka.testkit.{ ImplicitSender, TestKit }
 import org.scalatest.FunSuiteLike
 import org.scalatest.Matchers
 import scala.annotation.tailrec
