+++ akka-docs/rst/scala/code/docs/actor/FaultHandlingDocSpec.scala
@@ -10,8 +10,7 @@ import FaultHandlingDocSpec._
 //#testkit
 import com.typesafe.config.{ Config, ConfigFactory }
 import org.scalatest.{ FlatSpecLike, Matchers, BeforeAndAfterAll }
-import akka.testkit.scaladsl.TestKit
-import akka.testkit.{ ImplicitSender, TestActors, EventFilter }
+import akka.testkit.{ TestActors, TestKit, ImplicitSender, EventFilter }
 
 //#testkit
 object FaultHandlingDocSpec {
