+++ akka-docs/rst/scala/code/docs/testkit/TestKitUsageSpec.scala
@@ -3,8 +3,6 @@
  */
 package docs.testkit
 
-import akka.testkit.{ ImplicitSender, DefaultTimeout, TestActors }
-
 import language.postfixOps
 
 //#testkit-usage
@@ -20,7 +18,7 @@ import akka.actor.Actor
 import akka.actor.ActorRef
 import akka.actor.ActorSystem
 import akka.actor.Props
-import akka.testkit.scaladsl.TestKit
+import akka.testkit.{ TestActors, DefaultTimeout, ImplicitSender, TestKit }
 import scala.concurrent.duration._
 import scala.collection.immutable
