+++ akka-remote/src/test/scala/akka/remote/artery/compress/HandshakeShouldDropCompressionTableSpec.scala
@@ -119,7 +119,7 @@ class HandshakeShouldDropCompressionTableSpec extends ArteryMultiNodeSpec(Handsh
     }
   }
 
-  def waitForEcho(probe: TestKitBase, m: String, max: Duration = 3.seconds): Any =
+  def waitForEcho(probe: TestKit, m: String, max: Duration = 3.seconds): Any =
     probe.fishForMessage(max = max, hint = s"waiting for '$m'") {
       case `m` ⇒ true
       case x   ⇒ false
